show defect

+++ 服务/utils/Zip.py
@@ -1,44 +1,50 @@
 import os
 import time
 from pathlib import Path
-from threading import Thread
 from PIL import Image
-import shutil
 import numpy as np
 
+import Globs
 
-class ZipAndDeletion(Thread):
-    def __init__(self, path: Path):
-        self.path = path
+
+def _zip_(folder):
+    imageFolder = folder / "2D"
+    bmpList = list(imageFolder.glob("*.bmp"))
+    if not len(bmpList):
+        time.sleep(0.001)
+        return False
+    for imageUrl in bmpList:
+        image = Image.open(imageUrl)
+        image.save(imageUrl.with_suffix(".jpg"), quality=95, optimize=True)
+        image.close()
+        os.remove(str(imageUrl))
+    d3Folder = folder / "3D"
+    for d3Url in d3Folder.glob("*.npy"):
+        d3 = np.load(d3Url)
+        np.savez_compressed(d3Url.with_suffix(".npz"), array=d3)
+        os.remove(str(d3Url))
+    return True
+
+
+class ZipAndDeletion(Globs.ProcessClass):
+    def __init__(self, path: Path, reserve_num=0):
+        print(path)
+        self.path = Path(path)
+        self.reserve_num = reserve_num
         super().__init__()
 
     def run(self):
         folders = list(self.path.iterdir())
-        for folder in folders:
-            print(folder)
-            imageFolder = folder / "2D"
-            for imageUrl in imageFolder.glob("*.bmp"):
-                image=Image.open(imageUrl)
-                image.save(imageUrl.with_suffix(".jpg"),quality=95, optimize=True)
-                image.close()
-                os.remove(str(imageUrl))
-            d3Folder = folder / "3D"
-            for d3Url in d3Folder.glob("*.npy"):
-                d3 = np.load(d3Url)
-                np.savez_compressed(d3Url.with_suffix(".npz"),array=d3)
-                os.remove(str(d3Url))
-
-
-    def zip_bmp(self,bmp_url):
-        image = Image.open(bmp_url)
-        image.save(bmp_url.with_suffix(".jpg"))
-        image.close()
-        os.remove(str(bmp_url))
-
-    def zip_npy(self,npy_url):
-        npy = np.load(npy_url)
-        np.savez_compressed(npy_url.with_suffix(".npz"),array=npy)
-        os.remove(str(npy_url))
+        for folder in folders[self.reserve_num:][::-1]:
+            try:
+                sTime = time.time()
+                zipState = _zip_(folder)
+                eTime = time.time()
+                if zipState:
+                    print(f"{zipState} {folder} 数据压缩成功! 耗时: {eTime - sTime}")
+            except Exception as e:
+                print(f"数据压缩失败! {e}")
+        time.sleep(100)
 
     def delete(self):
         pass
